Refactor GraphQL schema
