+++ backend/server.js
@@ -1,10 +1,13 @@
-const { ApolloServer } = require('apollo-server');
-const typeDefs = require('./schema');
-const resolvers = require('./resolvers/resolvers.root');
+const { ApolloServer } = require("apollo-server");
+const typeDefs = require("./schema/schema.root");
+const resolvers = require("./resolvers/resolvers.root");
 
 // The ApolloServer constructor requires two parameters: your schema
 // definition and your set of resolvers.
-const server = new ApolloServer({ typeDefs, resolvers });
+const server = new ApolloServer({
+  typeDefs,
+  resolvers,
+});
 
 // The `listen` method launches a web server.
 server.listen().then(({ url }) => {
